pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apriflow-kubernetes"
version = "1.0.0"
dependencies = [
    "apache-airflow==2.10.4",
]
requires-python = ">=3.12"
authors = [
    { name = "Mustafa Abuelfadl", email = "mustafa@zidan.me" },
]

maintainers = [
    { name = "Mustafa Abuelfadl", email = "mustafa@zidan.me" },
]

description = "A Kubernetes example for Apache Airflow"
readme = {file = "README.md", content-type = "text/markdown"}
license = { file = "LICENSE.txt" }
keywords = ["apache-airflow", "kubernetes", "example"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
homepage = "https://medium.com/@moosezidan/introduction-to-running-airflow-on-kubernetes-2d9dc1625f7e"
Repository = "https://github.com/mustafa-zidan/airflow-kubernetes.git"

[tool.hatch.build.targets.wheel]
packages = ["src/airflow_kubernetes"]


